Added FlyWay

+++ src/main/java/com/example/model/order/entity/CustomerInfo.java
@@ -1,4 +1,4 @@
 package com.example.model.order.entity;
 
-public record CustomerInfo(Long id, String lastname, String firstname, double revenue) {
+public record CustomerInfo(Integer id, String lastname, String firstname, double revenue) {
 }
